Cleanup ruby.
ruby-2.0.0_p648 needs the backported rand_egd patch and the ssl3 patch.
ruby-2.1.8+ needs the backported rand_egd patch.
ruby-2.2.4+ needs the backported rand_egd patch but it's already in
ruby-patches.
ruby-2.3.0+ contains both the ssl3 and the randegd patch.

Instead of deleting the rand_egd methods, use a backported version of the patch
in ruby-2.3.0